+++ packages/builder/src/components/common/bindings/SnippetDrawer.svelte
